Add files via upload
For understanding the stress, we need to calculate the area on which the pressure is working. However for Pt nanoparticle because of its weird shape, its hard to decide which area we will consider. Thus here, I have calculated areas on top of Pt nanoparticle and also in the middle. Because of the nature of nanoparticles, we don't have a perfect plane like a macro materials, thus we have to be concern about how are we choosing plane. One single atom could be so much high, but we can't include that in the area calculation.